Add basic checking of NetworkConnection

+++ app/src/main/java/com/handysparksoft/trackmap/core/data/server/ServerDataSource.kt
@@ -2,16 +2,21 @@ package com.handysparksoft.trackmap.core.data.server
 
 import com.handysparksoft.data.source.RemoteDataSource
 import com.handysparksoft.domain.model.*
+import com.handysparksoft.trackmap.core.platform.ConnectionHandler
 
 class ServerDataSource(private val service: TrackMapService) : RemoteDataSource {
 
     override suspend fun saveUser(userId: String) {
+        if (ConnectionHandler.isNetworkConnected) {
         val user = User(userId, System.currentTimeMillis())
         service.saveUser(userId, user)
+        }
     }
 
     override suspend fun updateUser(userId: String, userProfileData: UserProfileData) {
-        service.updateUserName(userId, userProfileData)
+        if (ConnectionHandler.isNetworkConnected) {
+            service.updateUserName(userId, userProfileData)
+        }
     }
 
     override suspend fun getUserTrackMaps(userId: String): Map<String, TrackMap> {
@@ -23,23 +28,34 @@ class ServerDataSource(private val service: TrackMapService) : RemoteDataSource 
     }
 
     override suspend fun saveTrackMap(trackMapId: String, trackMap: TrackMap) {
-        service.saveTrackMap(trackMapId, trackMap)
+        if (ConnectionHandler.isNetworkConnected) {
+            service.saveTrackMap(trackMapId, trackMap)
+        }
     }
 
     override suspend fun saveUserTrackMap(userId: String, trackMapId: String, trackMap: TrackMap) {
-        service.saveUserTrackMap(userId, trackMapId, trackMap)
+        if (ConnectionHandler.isNetworkConnected) {
+            service.saveUserTrackMap(userId, trackMapId, trackMap)
+        }
     }
 
     override suspend fun joinTrackMap(userId: String, trackMapId: String, trackMapParticipant: TrackMapParticipant) {
-        service.joinTrackMap(trackMapId, trackMapParticipant)
+        if (ConnectionHandler.isNetworkConnected) {
+            service.joinTrackMap(trackMapId, trackMapParticipant)
+        }
     }
 
     override suspend fun getTrackMapById(trackMapId: String): TrackMap? {
-        return service.getTrackMaps().values.firstOrNull { it.trackMapId == trackMapId}
+        if (ConnectionHandler.isNetworkConnected) {
+            return service.getTrackMaps().values.firstOrNull { it.trackMapId == trackMapId }
+        }
+        return null
     }
 
     override suspend fun updateUserLocation(userId: String, latitude: Double, longitude: Double) {
-        service.updateUserLocation(userId, UserLocationData(latitude, longitude))
+        if (ConnectionHandler.isNetworkConnected) {
+            service.updateUserLocation(userId, UserLocationData(latitude, longitude))
+        }
     }
 
     override suspend fun leaveTrackMap(
@@ -47,7 +63,9 @@ class ServerDataSource(private val service: TrackMapService) : RemoteDataSource 
         trackMapId: String,
         trackMapParticipant: TrackMapParticipant
     ) {
-        service.deleteUserTrackMap(userId, trackMapId)
-        service.deleteTrackMapParticipantId(trackMapId, trackMapParticipant)
+        if (ConnectionHandler.isNetworkConnected) {
+            service.deleteUserTrackMap(userId, trackMapId)
+            service.deleteTrackMapParticipantId(trackMapId, trackMapParticipant)
+        }
     }
 }
